cache-contents-functions: refactored; introduced contents_index_all.gz
I ended up rewriting the code.  What was wrong with it?  I think it had
some problems with "Structured programming", or simply modular approach.
For example, previous code used to check "$no_contents_indices" variable
that comes from hsh-initroot.  This is wrong, from a design point of view.
"$no_contents_indices", despite being a global variable, should make sense
only in hsh-initroot (sadly enough, we cannot make it invisible for
hsh-sh-cache-contents-functions).  If "$no_contents_indices" is set in
hsh-initroot, it SIMPLY SHOULD NOT CALL any cache-contents functions.

On the other hand, hsh-initroot should not know too much about cache-contents
internals.  Why should it ever create cache/contents/{list,index}{,.new}
directories?  They are implementation details of the underlying "module".

That said, I provide two high-level "entry points" to cache-contents module.
Those are create_contents() and update_contents() functions, and their
semantics is just "do whatever is needed" (the former always rebuilds
the contents form scratch, while the latter possibly reuses the cache).
The only "contract" that hsh-sh-cache-contents-functions expects from
hsh-initroot is cache/contents directory set up.  The only "output",
after either of these two functions is called, is contents_index_bin
and contents_index_all global variables that point to corresponding
files under chroot/.host/.  That's it.

On the contents_index_all.

Why do we need it?  This is wrong place to answer.  In short, I want
to be able to resolve any pathname, to just standard PATH entries.
I also assume that new rpm-build algorithm will consider c_i_bin
to be "cheap", and "c_i_all" to be "expensive".

Why should it be compressed?  To save some disk space, and also memory
usage (because we build on tmpfs).  The compression is very efficient,
typical numbers are 90M -> 7M.  Decompression takes 0.5s.

One problem is that old rpm-build does not expect contents_index_all
to be compressed.  However, because gzipped data are "very binary", it
simply will not produce any result.

And now to some gory details.

When archiving the cache, blocking the signals is probably not the best
thing we want to do.  The idea is that we want "atomic transition", that
is, either the cache is archived completely, or nothing is archived at all.
Note, however, that we do not archive just a single file, but also the list/
and index/ state.  While mv(1) is usually "more atomic" than install(1),
there's still a chance that the sequence "copy_data+rotate_state" is not
atomic.  I change this to "remove_state+copy_data+rotate_state" sequence.
When something goes wrong after remove_state, we just lose the cache (which
is okay).  New cache is logically enabled only after the final rotate_state
call, which is "almost atomic" (for sure, I move the list/ AFTER the index/).

On the "install ... && Verbose ... || Fatal ..." clauses.  I think this
is overkill.  There's no reason why this install should normally fail.
And if it fails because of some "really bad" global problem, like disk
usage or malloc failure, we expect him say something useful before doing so.
And since we're in "sh -e" mode, that's probably okay.  We can't check just
everything, for we have to trust something.  What we do NOT trust is external
input data, which we must validate before using.

On the "if shell_function ...; then ...' clauses.  This is evil.  Don't use
shell function in the "if" conditions, because "sh -e" mode is disabled for
such calls.  You can no longer expect that something "just fails".

On the "cat file |while read ...; grep || echo >bad; .. done' clauses.
I replaced them with "while ... done <file".