hsh-rebuild --query-repackage: re-use make_srpm_from_pkgtar()
The drawback of this unification is that the user is switched more
times: beside all the "foreign" calls in make_srpm_from_pkgtar() --
namely 3, there is 1 more: to unpack the srpm.

Before this change, there was 1 such call for this operation, because
the initial actions and the final actions were glued to the main
actions. Now, when a separate function is used, it's not so simple
to optimize like this.