Robustify use of exclude_docs configuration variable
exclude_docs configuration variable can be used to achieve the effect of
--excludedocs option in hsh, hsh-initroot, hsh-install, and hsh-rebuild.

When exclude_docs had a non-empty value, that value was assumed to be
set to --excludedocs, and it used to be passed to rpm without any
further tampering, surprising many unaware users.

Fix this issue by following the KISS principle: instead of passing the
value of exclude_docs to rpm, pass --excludedocs when exclude_docs
variable has a non-empty value.

The change itself is the result of the following command:

git grep -Fl '$exclude_docs' -- hasher/hsh-* |
	xargs sed -i 's/\$exclude_docs/${exclude_docs:+--excludedocs}/g'

Reported-by: Alexey Gladkov <legion@altlinux.org>