Introduce --rpmbuild-args for hsh(1) and hsh-rebuild(1)
Both these utilities previously supported passing additional arguments
to rpmbuild during package build orchestration.  However, one of them
used the option --build-args='<partial-command-line>', the other used
the option --args='<partial-command-line>', resulting in endless
confusion.

We introduce a unified synonym for those options and deprecate them.

The very name --rpmbuild-args was chosen to be explicit but concise;
gear(1) already sets a standard for this kind of CLI passthrough
mark-up, as it can be seen in the output of `gear --help'.  Moreover,
this option name fits very well into the possible future scenario where
hasher learns to deal with a non-RPM packaging ecosystem with a
different, independent CLI from rpm and rpmbuild.  The increased command
length is not much of a concern, since the excess typing can be
alleviated with shell autocompletion.