Fix optional parameters to CLI options
A call to getopt(1) changes the semantics of a command line if an option
has a non-required argument: --key[=value].

Prior to this, the systemd-run wrapper was hooked after the getopt call
and thus tried to re-execute a broken command line. For example, the
following command:

  `hsh-shell --shell=/bin/bash workdir`

was turned into:

  `hsh-shell --shell /bin/bash workdir`,

silently failing and confusing the users.

Fixes: fa974b55d40c6b5536e5323ab08c76ed5347952e