create_entry_header: source /etc/default-environment in the chroot
While fulfilling a chrootuid request, hasher-privd(8) passes the same
hardcoded PATH value to anything it spawns, in any chroot, for any
privilege triplet. This is an issue since we want different PATH values
for chroots based on different ALT repository branches. If we change
the hardcoded PATH in hasher-priv, it will not help. We need a way to
take the PATH value for contained processes from inside the chroot.

The culprit here is that PATH assignment is done in /etc/profile, which
is a script with overloaded functionality. It is intended to be read by
a particular interpreter and does its job through manipulating that
interpreter's internal state, as opposed to a self-contained approach.
It is also deeply coupled with the job of a login shell, since that
appeared to make sense historically.

In the future that file might be replaced with a maintainable mechanism;
bonus points if that would be reusable by our rc systems.